refactor core error modules
